requirements.txt
## The following requirements were added by pip freeze:
adal==1.2.1
aiohttp==3.5.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
azure-common==1.1.19
azure-cosmos==3.0.2
azure-eventgrid==1.2.0
azure-nspkg==3.0.2
azure-storage-common==2.0.0
azure-storage-queue==2.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
discord.py==1.1.1
idna==2.8
idna-ssl==1.1.0
isodate==0.6.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.6.6
msrestazure==0.6.0
multidict==4.5.2
mysqlclient==1.4.2.post1
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
websockets==6.0
wrapt==1.11.1
yarl==1.3.0
